Numerous changes. See full commit message.
Added: isCountable() (countable()) and isArray for determining if options should be countable options or treated as arrays.
Added: filter() as an alias to test().
Removed: "value" default; only "optionalValue" is used in conjunction with "boolean" to determine if a value is expected or not.
Changed: Anything calling "value," "optionalValue," or "boolean" internally has had its behavior altered.
Changed: All options are now treated as expecting optional values UNLESS a boolean() OR empty value() call is made.
Fixed: "filterValue" was referred to elsewhere in the code as "filterFailure."